% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DS_schema_cols.R
\name{DS_schema_cols}
\alias{DS_schema_cols}
\title{Available and required columns in the DataStream Schema}
\usage{
DS_schema_cols()
}
\value{
a list of available and required column names
}
\description{
Available and required columns in the DataStream Schema
}
\examples{
library(datastreamr.helper)
DS_schema_cols()
}
