% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DS_valid_vals.R
\name{DS_valid_vals}
\alias{DS_valid_vals}
\title{Valid fields for DataStream API queries}
\usage{
DS_valid_vals()
}
\value{
a list of valid fields for \code{select} and \code{filter}
}
\description{
Valid fields for DataStream API queries
}
\examples{
library(datastreamr.helper)
DS_valid <- DS_valid_vals()

# Valid fields to select
DS_valid$valid_select

# Valid fields to filter
DS_valid$valid_filter

# valid values for lookups
DS_valid$valid_lookup$ActivityMediaName
}
