% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DS_validator.R
\name{DS_validator}
\alias{DS_validator}
\title{DataStream Schema Validator}
\usage{
DS_validator(x, .chunk_size = 250L)
}
\arguments{
\item{x}{a data frame or file path to a CSV to be uploaded to DataStream}

\item{.chunk_size}{an integer indicating the number of rows to read into R at a time}
}
\value{
a data frame of error messages from schema validation
}
\description{
Validate supplied data frame against DataStream Schema: https://datastream.org/schema
}
\details{
Schema checks related to pattern matching are currently unavailable. Function will save data frame to a temporary file internally.
}
\examples{
dt <- tibble::tribble(
~DatasetName, ~MonitoringLocationID, ~MonitoringLocationName, ~MonitoringLocationLatitude, ~MonitoringLocationLongitude, ~MonitoringLocationHorizontalCoordinateReferenceSystem, ~MonitoringLocationHorizontalAccuracyMeasure, ~MonitoringLocationHorizontalAccuracyUnit, ~MonitoringLocationVerticalMeasure, ~MonitoringLocationVerticalUnit, ~MonitoringLocationType,   ~ActivityType, ~ActivityMediaName, ~ActivityStartDate, ~ActivityStartTime, ~ActivityEndDate, ~ActivityEndTime, ~ActivityDepthHeightMeasure, ~ActivityDepthHeightUnit, ~SampleCollectionEquipmentName,  ~CharacteristicName, ~MethodSpeciation, ~ResultSampleFraction, ~ResultValue, ~ResultUnit, ~ResultValueType, ~ResultDetectionCondition, ~ResultDetectionQuantitationLimitMeasure, ~ResultDetectionQuantitationLimitUnit, ~ResultDetectionQuantitationLimitType, ~ResultStatusID, ~ResultComment, ~ResultAnalyticalMethodID, ~ResultAnalyticalMethodContext, ~ResultAnalyticalMethodName, ~AnalysisStartDate, ~AnalysisStartTime, ~AnalysisStartTimeZone, ~LaboratoryName, ~LaboratorySampleID,
"test",             "test-314",               "test-314",                  43.5895361,                  -79.9411775,                                                "NAD83",                                           NA,                                        NA,                                 NA,                              NA,          "River/Stream", "Field Msr/Obs",    "Surface Water",       "2005-07-12",                 NA,               NA,               NA,                          NA,                       NA,                 "Probe/Sensor", "Temperature, water",                NA,                    NA,           18,     "deg C",         "Actual",                        NA,                                       NA,                                    NA,                                    NA,              NA,             NA,                        NA,                             NA,                          NA,                 NA,                 NA,                     NA,              NA,                  NA
)

DS_validator(dt)

dt_bad <- dt
dt_bad$CharacteristicName <- "Something Wrong"

dt_new <- rbind(dt,dt_bad)

DS_validator(dt_new)

}
