% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qs_helper.R
\name{qs_helper}
\alias{qs_helper}
\title{Format inputs for \emph{datastreamr}}
\usage{
qs_helper(select = NULL, filter = NULL, top = NULL, count = NULL)
}
\arguments{
\item{select}{a vector of fields to request from the API.}

\item{filter}{a named list of fields to filter by from the API. Each element in the list should be a list or vector in which the first element represents a R operator ('==','\%in\%','>=', etc.), and subsequent elements representing the criteria.}

\item{top}{an integer representing the number of rows query at a time (max = 10000).}

\item{count}{a logical value indicating whether to return the number of observations rather than the data itself.}
}
\value{
a list formatted for input as the \code{qs} argument (i.e., \link[datastreamr]{metadata})
}
\description{
For details see: \href{v}{datastreamr}
}
\examples{

library(datastrear.helper)
qs_helper(
  select=c("ActivityStartDate",
           "ActivityStartTime",
           "CharacteristicName",
           "ResultUnit",
           "ResultValue",
           "ResultDetectionCondition",
           "ResultStatusID",
           "ResultComment",
           "CreateTimestamp"),
  filter=list(
           DOI=c("==","10.25976/0gvo-9d12"),
           CharacteristicName=c("\%in\%",c("Specific conductance","pH")),
           ActivityStartYear=c(">",2010),
           LocationId=c("==","862774")
           ),
  top=5000L
  )

}
